+++ frontend/src/app/map/page.tsx
@@ -1,0 +1,7 @@
+'use client'
+import MapPage from '@/features/map/MapPage'
+import React from 'react'
+
+export default function page() {
+  return <MapPage />
+}
