--- frontend/src/app/spot/page.tsx
@@ -1,6 +1,0 @@
-import SpotPage from '@/features/spot/SpotPage'
-import React from 'react'
-
-export default function page() {
-  return <SpotPage />
-}
