--- frontend/src/features/spot/SpotPage.tsx
@@ -1,22 +1,0 @@
-'use client'
-
-import React from 'react'
-import { APIProvider, Map } from '@vis.gl/react-google-maps'
-
-export default function SpotPage() {
-  const API_KEY = process.env.NEXT_PUBLIC_GOOGLE_MAPS_API_KEY as string
-
-  return (
-    <div>
-      <APIProvider apiKey={API_KEY}>
-        <Map
-          style={{ width: '100vw', height: '100vh' }}
-          defaultCenter={{ lat: 22.54992, lng: 0 }}
-          defaultZoom={3}
-          gestureHandling={'greedy'}
-          disableDefaultUI={true}
-        />
-      </APIProvider>
-    </div>
-  )
-}
